requirements.txt
contractions>=0.1.73
fastapi>=0.115.0
pydantic>=2.9.0
pyspellchecker>=0.8.1
pytest>=8.3.0
httpx>=0.27.0
spacy>=3.8.0
uvicorn>=0.30.0
